Fix unsound problem in `Row::row_unchecked` (#5)
Hi, thanks for your folk of arrow and best wishes first.

We have found a simple unsound problem in `::compute::sort::row::Rows`.
The function
[`Rows::row_unchecked`](https://github.com/rerun-io/re_arrow2/blob/33a32000001df800e4840d92c33b03e7007311e1/src/compute/sort/row/mod.rs#L272)
is wrongly marked as safe, which would confuse the boundary between safe
and unsafe Rust.

Since the repository of arrow2 have been archived, we cannot create any
issue report. So we directly open this PR to mark the function
`row_unchecked` as 'unsafe' and add a `debug_assert!` macro for safety
consideration.